Issue #16 refactor: AddSchedule -> SaveSchedule 네이밍 변경
Schedule 추가 및 생성 두 가지 동작을 하기 때문에 Save가 더 적절한 단어인 듯 하여
네이밍을 변경함.
