--- app/src/main/java/com/drunkenboys/calendarun/ui/addschedule/model/BehaviorType.kt
@@ -1,5 +1,0 @@
-package com.drunkenboys.calendarun.ui.addschedule.model
-
-enum class BehaviorType {
-    INSERT, UPDATE
-}
