+++ app/src/main/java/com/drunkenboys/calendarun/ui/saveschedule/model/BehaviorType.kt
@@ -1,0 +1,5 @@
+package com.drunkenboys.calendarun.ui.saveschedule.model
+
+enum class BehaviorType {
+    INSERT, UPDATE
+}
